requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.12.1
amqp==2.6.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
arrow==0.13.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
billiard==3.6.4.0
blinker==1.7.0
celery==4.3.0
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
click==7.1.2
cpe==1.2.1
cryptography==41.0.5
cssmin==0.2.0
deepdiff==4.0.7
Deprecated==1.2.14
distro==1.8.0
dnspython==2.4.2
email-validator==1.1.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Flask==1.0.2
Flask-Admin==1.5.5
Flask-Assets==0.12
Flask-DebugToolbar==0.10.1
Flask-Gravatar==0.5.0
Flask-Limiter==1.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
flask-paginate==0.5.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.3.2
Flask-User==1.0.2.2
Flask-WTF==0.14.3
frozenlist==1.4.0
gunicorn==20.0.4
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==2.10
importlib-metadata==4.13.0
importlib-resources==6.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsmin==3.0.1
jsonpickle==3.0.2
kombu==4.6.11
limits==3.7.0
Mako==1.3.0
MarkupSafe==1.1.1
multidict==6.0.4
nested-lookup==0.2.18
openai==0.27.0
opencve==1.4.1
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.2
passlib==1.7.4
psycopg2-binary==2.8.4
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
pytz==2023.3.post1
redis==3.3.11
requests==2.23.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.3.1
SQLAlchemy-Utils==0.33.11
tqdm==4.66.1
types-pytz==2023.3.1.1
typing_extensions==4.8.0
tzdata==2023.3
untangle==1.1.1
urllib3==1.25.11
vine==1.3.0
webassets==2.0
Werkzeug==1.0.1
wrapt==1.16.0
WTForms==3.1.1
yarl==1.9.3
zipp==3.17.0
zhipuai==2.0.1.20240429